# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Email Marketing Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  pending: "pending",
  sent: "sent",
  failed: "failed",
} as const;

export interface Lead {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  company: string | null;
  /** @nullable */
  source: string | null;
  status: LeadStatus;
  /** @nullable */
  emailSentAt: string | null;
  /** @nullable */
  errorMessage: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface LeadsPage {
  leads: Lead[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateLeadBody {
  name: string;
  email: string;
  phone?: string;
  company?: string;
  source?: string;
}

export type UpdateLeadBodyStatus =
  (typeof UpdateLeadBodyStatus)[keyof typeof UpdateLeadBodyStatus];

export const UpdateLeadBodyStatus = {
  pending: "pending",
  sent: "sent",
  failed: "failed",
} as const;

export interface UpdateLeadBody {
  name?: string;
  email?: string;
  phone?: string;
  company?: string;
  source?: string;
  status?: UpdateLeadBodyStatus;
}

export interface ImportLeadsBody {
  leads: CreateLeadBody[];
}

export interface ImportResult {
  inserted: number;
  skipped: number;
  errors: string[];
}

export interface EmailTemplate {
  id: number;
  subject: string;
  body: string;
  updatedAt: string;
}

export interface SaveTemplateBody {
  subject: string;
  body: string;
}

export interface SendResult {
  sent: number;
  failed: number;
  message: string;
}

export type EmailLogStatus =
  (typeof EmailLogStatus)[keyof typeof EmailLogStatus];

export const EmailLogStatus = {
  sent: "sent",
  failed: "failed",
} as const;

export interface EmailLog {
  id: number;
  leadId: number;
  leadName: string;
  leadEmail: string;
  /** @nullable */
  company: string | null;
  subject: string;
  status: EmailLogStatus;
  /** @nullable */
  errorMessage: string | null;
  sentAt: string;
}

export interface EmailLogsPage {
  logs: EmailLog[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface DashboardStats {
  totalLeads: number;
  emailsSent: number;
  emailsPending: number;
  emailsFailed: number;
  successRate: number;
  recentlySent: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  email_sent: "email_sent",
  email_failed: "email_failed",
  lead_added: "lead_added",
  lead_imported: "lead_imported",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  /** @nullable */
  leadName: string | null;
  /** @nullable */
  leadEmail: string | null;
  timestamp: string;
}

export type GetLeadsParams = {
  search?: string;
  status?: GetLeadsStatus;
  page?: number;
  limit?: number;
};

export type GetLeadsStatus =
  (typeof GetLeadsStatus)[keyof typeof GetLeadsStatus];

export const GetLeadsStatus = {
  pending: "pending",
  sent: "sent",
  failed: "failed",
  all: "all",
} as const;

export type GetEmailLogsParams = {
  page?: number;
  limit?: number;
};

export type GetDashboardActivityParams = {
  limit?: number;
};