# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Email Marketing Automation API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead management
  - name: templates
    description: Email template management
  - name: emails
    description: Email sending and tracking
  - name: dashboard
    description: Dashboard stats and activity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads with search, filter, and pagination
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, sent, failed, all]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsPage"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads/import:
    post:
      operationId: importLeads
      tags: [leads]
      summary: Bulk import leads from parsed Excel/CSV data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportLeadsBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /templates:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get the global email template
      responses:
        "200":
          description: Email template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    put:
      operationId: saveTemplate
      tags: [templates]
      summary: Save or update the global email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTemplateBody"
      responses:
        "200":
          description: Saved template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"

  /emails/send:
    post:
      operationId: sendPendingEmails
      tags: [emails]
      summary: Send emails to all pending leads
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"

  /emails/send/{leadId}:
    post:
      operationId: sendEmailToLead
      tags: [emails]
      summary: Send email to a specific lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emails/retry:
    post:
      operationId: retryFailedEmails
      tags: [emails]
      summary: Retry all failed emails
      responses:
        "200":
          description: Retry result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"

  /emails/retry/{leadId}:
    post:
      operationId: retryEmailForLead
      tags: [emails]
      summary: Retry failed email for a specific lead
      parameters:
        - name: leadId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retry result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /emails/logs:
    get:
      operationId: getEmailLogs
      tags: [emails]
      summary: Get recent email send logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Email send logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLogsPage"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, sent, failed]
        emailSentAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - company
        - source
        - status
        - emailSentAt
        - errorMessage
        - createdAt
        - updatedAt

    LeadsPage:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - leads
        - total
        - page
        - limit
        - totalPages

    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        source:
          type: string
      required:
        - name
        - email

    UpdateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        source:
          type: string
        status:
          type: string
          enum: [pending, sent, failed]

    ImportLeadsBody:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/CreateLeadBody"
      required:
        - leads

    ImportResult:
      type: object
      properties:
        inserted:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - inserted
        - skipped
        - errors

    EmailTemplate:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        body:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - subject
        - body
        - updatedAt

    SaveTemplateBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
      required:
        - subject
        - body

    SendResult:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        message:
          type: string
      required:
        - sent
        - failed
        - message

    EmailLog:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        leadName:
          type: string
        leadEmail:
          type: string
        company:
          type: ["string", "null"]
        subject:
          type: string
        status:
          type: string
          enum: [sent, failed]
        errorMessage:
          type: ["string", "null"]
        sentAt:
          type: string
      required:
        - id
        - leadId
        - leadName
        - leadEmail
        - company
        - subject
        - status
        - errorMessage
        - sentAt

    EmailLogsPage:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/EmailLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - logs
        - total
        - page
        - limit
        - totalPages

    DashboardStats:
      type: object
      properties:
        totalLeads:
          type: integer
        emailsSent:
          type: integer
        emailsPending:
          type: integer
        emailsFailed:
          type: integer
        successRate:
          type: number
        recentlySent:
          type: integer
      required:
        - totalLeads
        - emailsSent
        - emailsPending
        - emailsFailed
        - successRate
        - recentlySent

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [email_sent, email_failed, lead_added, lead_imported]
        message:
          type: string
        leadName:
          type: ["string", "null"]
        leadEmail:
          type: ["string", "null"]
        timestamp:
          type: string
      required:
        - id
        - type
        - message
        - leadName
        - leadEmail
        - timestamp