# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Email Marketing Automation API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List leads with search, filter, and pagination
 */
export const getLeadsQueryPageDefault = 1;
export const getLeadsQueryLimitDefault = 20;

export const GetLeadsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  status: zod.enum(["pending", "sent", "failed", "all"]).optional(),
  page: zod.coerce.number().default(getLeadsQueryPageDefault),
  limit: zod.coerce.number().default(getLeadsQueryLimitDefault),
});

export const GetLeadsResponse = zod.object({
  leads: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      email: zod.string(),
      phone: zod.string().nullable(),
      company: zod.string().nullable(),
      source: zod.string().nullable(),
      status: zod.enum(["pending", "sent", "failed"]),
      emailSentAt: zod.string().nullable(),
      errorMessage: zod.string().nullable(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Create a new lead
 */
export const CreateLeadBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().optional(),
  company: zod.string().optional(),
  source: zod.string().optional(),
});

/**
 * @summary Bulk import leads from parsed Excel/CSV data
 */
export const ImportLeadsBody = zod.object({
  leads: zod.array(
    zod.object({
      name: zod.string(),
      email: zod.string(),
      phone: zod.string().optional(),
      company: zod.string().optional(),
      source: zod.string().optional(),
    }),
  ),
});

export const ImportLeadsResponse = zod.object({
  inserted: zod.number(),
  skipped: zod.number(),
  errors: zod.array(zod.string()),
});

/**
 * @summary Update a lead
 */
export const UpdateLeadParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLeadBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  company: zod.string().optional(),
  source: zod.string().optional(),
  status: zod.enum(["pending", "sent", "failed"]).optional(),
});

export const UpdateLeadResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullable(),
  company: zod.string().nullable(),
  source: zod.string().nullable(),
  status: zod.enum(["pending", "sent", "failed"]),
  emailSentAt: zod.string().nullable(),
  errorMessage: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a lead
 */
export const DeleteLeadParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get the global email template
 */
export const GetTemplateResponse = zod.object({
  id: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Save or update the global email template
 */
export const SaveTemplateBody = zod.object({
  subject: zod.string(),
  body: zod.string(),
});

export const SaveTemplateResponse = zod.object({
  id: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Send emails to all pending leads
 */
export const SendPendingEmailsResponse = zod.object({
  sent: zod.number(),
  failed: zod.number(),
  message: zod.string(),
});

/**
 * @summary Send email to a specific lead
 */
export const SendEmailToLeadParams = zod.object({
  leadId: zod.coerce.number(),
});

export const SendEmailToLeadResponse = zod.object({
  sent: zod.number(),
  failed: zod.number(),
  message: zod.string(),
});

/**
 * @summary Retry all failed emails
 */
export const RetryFailedEmailsResponse = zod.object({
  sent: zod.number(),
  failed: zod.number(),
  message: zod.string(),
});

/**
 * @summary Retry failed email for a specific lead
 */
export const RetryEmailForLeadParams = zod.object({
  leadId: zod.coerce.number(),
});

export const RetryEmailForLeadResponse = zod.object({
  sent: zod.number(),
  failed: zod.number(),
  message: zod.string(),
});

/**
 * @summary Get recent email send logs
 */
export const getEmailLogsQueryPageDefault = 1;
export const getEmailLogsQueryLimitDefault = 20;

export const GetEmailLogsQueryParams = zod.object({
  page: zod.coerce.number().default(getEmailLogsQueryPageDefault),
  limit: zod.coerce.number().default(getEmailLogsQueryLimitDefault),
});

export const GetEmailLogsResponse = zod.object({
  logs: zod.array(
    zod.object({
      id: zod.number(),
      leadId: zod.number(),
      leadName: zod.string(),
      leadEmail: zod.string(),
      company: zod.string().nullable(),
      subject: zod.string(),
      status: zod.enum(["sent", "failed"]),
      errorMessage: zod.string().nullable(),
      sentAt: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalLeads: zod.number(),
  emailsSent: zod.number(),
  emailsPending: zod.number(),
  emailsFailed: zod.number(),
  successRate: zod.number(),
  recentlySent: zod.number(),
});

/**
 * @summary Get recent activity feed
 */
export const getDashboardActivityQueryLimitDefault = 10;

export const GetDashboardActivityQueryParams = zod.object({
  limit: zod.coerce.number().default(getDashboardActivityQueryLimitDefault),
});

export const GetDashboardActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["email_sent", "email_failed", "lead_added", "lead_imported"]),
  message: zod.string(),
  leadName: zod.string().nullable(),
  leadEmail: zod.string().nullable(),
  timestamp: zod.string(),
});
export const GetDashboardActivityResponse = zod.array(
  GetDashboardActivityResponseItem,
);